+++ src/js/helpers/account-helper.js
@@ -34,3 +34,11 @@ export function lockAccount (pair) {
   if (pair.meta.isTesting) return
   pair.lock()
 }
+
+export function cropAddress (string, reduceValue = 4) {
+  if (string.length > 2 * reduceValue) {
+    return `${string.slice(0, reduceValue)}…${string.slice(-reduceValue)}`
+  } else {
+    return string
+  }
+}