+++ CHANGELOG.md
@@ -45,6 +45,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Transfer form
 - ApiCaller
 - Account deposit on Account row
+- Status message on transfer form
+- Holders page
+- Blockchain events constants
+- Map page
 
 #### Changed
 - Folders structure
@@ -53,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `useCall`
   `vuex modules`
 - Icon on account row
+- Design status message
 
 #### Removed
 - Explorer helper
@@ -62,5 +67,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Initial project release
 - Initiated project with vue cli
 
-[Unreleased]: https://github.com/LedgerProject/LiV_frontend_2.0/compare/0.1.0-rc.1...main
-[0.1.0-rc.1]: https://gitlab.com/tokend/polkadot/web-client/tags/0.1.0-rc.1
+[Unreleased]: https://gitlab.com/tokend/polkadot/new-web-client/compare/0.1.0-rc.1...main
+[0.1.0-rc.1]: https://gitlab.com/tokend/polkadot/new-web-client/tags/0.1.0-rc.1